+++ src/styled/theme.js
@@ -1,0 +1,7 @@
+/* Theme */
+export const COLOR_PRIMARY = '#034263';
+export const COLOR_SECONDARY = '#BAE1FF';
+export const COLOR_LIGHT_GREY = '#EAEAEA';
+export const COLOR_GREY = '#c5c5c5';
+export const COLOR_DARK_GREY = '#7D7D7D';
+export const COLOR_WHITE = '#FFFFFF';
